🏗️ Architect: Modularized apts/plotting/path.py
Refactored the monolithic `apts/plotting/path.py` into a modular package `apts/plotting/path/`.
- Extracted shared logic (markers, hemisphere detection, cardinal markers) into `utils.py`.
- Moved Sun and Moon path plotting logic to `sun.py` and `moon.py` respectively.
- Maintained public API via `__init__.py`.
- Updated all internal imports and unit test mocks.
- Improved robustness of `moon_illumination` formatting to handle `MagicMock` in tests.
- Verified all tests pass and linting is clean.
