+++ tests/plotting/test_plot_path.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock, patch
 from apts.plotting import plot_sun_and_moon_path
 
 
-@patch("apts.plotting.path.generate_plot_sun_path")
+@patch("apts.plotting.path.sun.generate_plot_sun_path")
 def test_plot_sun_and_moon_path_sun(mock_generate_plot_sun_path):
     mock_observation = MagicMock()
     mock_observation.sun_observation = True
@@ -11,7 +11,7 @@ def test_plot_sun_and_moon_path_sun(mock_generate_plot_sun_path):
     mock_generate_plot_sun_path.assert_called_once_with(mock_observation.place, None)
 
 
-@patch("apts.plotting.path.generate_plot_moon_path")
+@patch("apts.plotting.path.moon.generate_plot_moon_path")
 def test_plot_sun_and_moon_path_moon(mock_generate_plot_moon_path):
     mock_observation = MagicMock()
     mock_observation.sun_observation = False
