+++ tests/unit/test_place.py
@@ -408,9 +408,9 @@ class TestPlacePlotting(unittest.TestCase):
         self.place._moon_phase_letter = MagicMock(return_value="M")
         self.place.moon_illumination = MagicMock(return_value=71)
 
-    @patch("apts.plotting.path.gettext_")
+    @patch("apts.plotting.path.moon.gettext_")
     @patch("apts.utils.plot.gettext_")
-    @patch("apts.plotting.path.get_dark_mode")  # Corrected path for get_dark_mode used in Place
+    @patch("apts.plotting.path.moon.get_dark_mode")  # Corrected path for get_dark_mode used in Place
     @patch("pandas.DataFrame.plot")
     def test_plot_moon_path_styling(
         self, mock_df_plot, mock_get_dark_mode_place, mock_gettext, mock_place_gettext
@@ -563,7 +563,7 @@ class TestPlacePlotting(unittest.TestCase):
                 mock_df_plot.call_count = 0
 
 
-@patch("apts.plotting.path.get_dark_mode")
+@patch("apts.plotting.path.moon.get_dark_mode")
 class TestDarkMode(unittest.TestCase):
     def setUp(self):
         self.place = setup_place()
@@ -633,7 +633,7 @@ class TestPlaceSouthernHemisphere(unittest.TestCase):
         # Sydney, Australia
         self.place = Place(lat=-33.8688, lon=151.2093)
 
-    @patch("apts.plotting.path.get_dark_mode", return_value=False)
+    @patch("apts.plotting.path.moon.get_dark_mode", return_value=False)
     def test_plot_moon_path_southern_hemisphere_non_mocked(self, mock_get_dark_mode):
         """
         Non-mocked test to verify the southern hemisphere moon path plot.
